requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
gunicorn==20.1.0
Pillow==9.5.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
SQLAlchemy==2.0.15
alembic==1.11.1
pytz==2023.3
python-dotenv==1.0.0
email-validator==2.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Babel==4.0.0
pymongo==4.0.0
openai>=1.0.0